Add destructor and reorgnise constructor
