+++ linalg/matrix.h
@@ -66,44 +66,25 @@ template<typename ITYPE, typename VTYPE>
 class sparse_matrix {
 
     public:
+        // *****************************************
+        // ************* Constructors **************
+        // *****************************************
+
         // Constructor - takes in size of row and column to generate an empty matrix
-        sparse_matrix<ITYPE,VTYPE>(ITYPE r, ITYPE c)
-            : rows(r), cols(c), dummy_(0) {}
+        sparse_matrix<ITYPE,VTYPE>(ITYPE, ITYPE);
+        // Constructor - null matrix
+        sparse_matrix<ITYPE,VTYPE>();
+        // Constructor - takes in another sparse_matrix object as an argument
+        sparse_matrix<ITYPE,VTYPE>(const sparse_matrix<ITYPE,VTYPE> &);
+        // Constructor - generate various square matrices
+        sparse_matrix<ITYPE,VTYPE>(ITYPE, const std::string & = {});
 
-        // Overload constructor - null
-        sparse_matrix<ITYPE,VTYPE>()
-            : rows(0), cols(0), dummy_(0) {}
+        // *****************************************
+        // ************* Destructor **************
+        // *****************************************
 
-        // Overload constructor - takes in another sparse_matrix object as an argument
-        sparse_matrix<ITYPE,VTYPE>(const sparse_matrix<ITYPE,VTYPE> &exist_mat)
-            : rows(exist_mat.rows), cols(exist_mat.cols), dummy_(0) {
-              // Copy the data from the existing matrix
-              std::vector<std::pair<ITYPE, ITYPE>> vec_temp;
-              vec_temp = exist_mat.get_non_zero_elements();
-              for (auto elememt : vec_temp) {
-                  if (std::abs(exist_mat.get_value(elememt.first, elememt.second)) >= std::numeric_limits<VTYPE>::min())
-                      matrix[std::make_pair(elememt.first, elememt.second)] = exist_mat.get_value(elememt.first, elememt.second);
-              }
-          }
-
-        // Overload constructor - generate various square matrix
-        sparse_matrix<ITYPE,VTYPE>(ITYPE n, const std::string &token = {})
-            : rows(n), cols(n), dummy_(0) {
-            if(token.empty()) {
-                // empty (all-zero) square matrix (Do nothing from the code perspective)
-            } else if(trim(token)=="identity") {
-                // identity square matrix
-                for (ITYPE i = 0; i < n; ++i) {
-                    matrix[std::make_pair(i, i)] = static_cast<VTYPE>(1);
-                }
-            } else {
-                std::cerr << "MUI Error [matrix.h]: unidentified token string for square matrix constructor" << std::endl;
-                std::cerr << "    Please set the token string as:" << std::endl;
-                std::cerr << "    empty string (default): Empty (all-zero) square matrix" << std::endl;
-                std::cerr << "    'identity': identity square matrix" << std::endl;
-                std::abort();
-            }
-        }
+        // Destructor
+        ~sparse_matrix<ITYPE,VTYPE>();
 
         // Member function to print matrix elements to the console
         void print() {
@@ -352,11 +333,13 @@ class sparse_matrix {
         sparse_matrix<ITYPE,VTYPE> transpose();
 
     private:
+        // Non-zero sparse matrix elements in COO format
         std::map<std::pair<ITYPE, ITYPE>, VTYPE> matrix;
+        // Number of rows of sparse matrix
         ITYPE rows;
+        // Number of columns of sparse matrix
         ITYPE cols;
-
-        // Dummy member variable for invalid or unassigned element in the matrix
+        // Dummy member variable for invalid or unassigned elements in sparse matrix
         VTYPE dummy_;
 };
 
@@ -364,6 +347,7 @@ class sparse_matrix {
 } // mui
 
 // Include implementations
+#include "matrix_ctor_dtor.h"
 #include "matrix_arithmetic.h"
 
 #endif /* MUI_SPARSE_MATRIX_H_ */
